requirements.txt
tqdm
numpy
scikit-learn
pytorch
gensim
pickle